requirements.txt
mysql-connector==2.2.9
pandas==2.1.4
pathlib==1.0.1
pymysql==1.1.1
requests==2.31.0
sqlalchemy==2.0.25
